refactor: extract game card into shared component
